Android: Support interchangeable BIOS files
The previous code supported replacing the BIOS with a regional counterpart, but destructively replaced the existing image with the new one. This should allow swapping to a new region without removing that region as a future option, so it can be switched between multiple regions with multiple files. It will still fall back to the previous method simply to prevent the removal of a useable BIOS.